server/requirements.txt
openenv-core>=0.1.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
requests>=2.31.0
openai>=1.1.0
